# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SECOM MES Database API",
    "description": "REST API for SECOM Manufacturing Execution System Database\n\nThis API provides access to semiconductor manufacturing data including:\n- Master data (Equipment, Shifts, Operators, Product Types, Features)\n- Production data (Lots, Measurements, Quality Results)\n- Analytics (Production summaries, Equipment health, Shift performance, Quality analytics)\n\n**Database**: MariaDB 11.2\n**Total Records**: 1,567 lots with 590 sensor measurements each\n**Time Period**: July 2008 - September 2008\n\n**Key Endpoints**:\n- `/equipment` - Manufacturing equipment management\n- `/lots` - Production lot tracking with filters\n- `/measurements` - Sensor measurement data\n- `/quality` - Quality inspection results\n- `/analytics` - Analytics and reporting endpoints\n",
    "contact": {
      "name": "SECOM MES Team",
      "email": "mes@secom.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/features/{featureId}": {
      "get": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Get feature by ID",
        "description": "Returns a single feature metadata by feature ID",
        "operationId": "getFeatureById",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureMeta"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Update feature",
        "description": "Updates an existing feature metadata record",
        "operationId": "updateFeature",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureMeta"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureMeta"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Delete feature",
        "description": "Deletes a feature metadata record",
        "operationId": "deleteFeature",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/features": {
      "get": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Get all features",
        "description": "Returns paginated list of all feature metadata (590 features)",
        "operationId": "getAllFeatures",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageFeatureMeta"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Create new feature",
        "description": "Creates a new feature metadata record",
        "operationId": "createFeature",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureMeta"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureMeta"
                }
              }
            }
          }
        }
      }
    },
...
    "/measurements/feature/{featureId}": {
      "get": {
        "tags": [
          "Measurements"
        ],
        "summary": "Get measurements for a feature",
        "description": "Returns all measurements for a specific sensor/feature across all lots",
        "operationId": "getMeasurementsByFeature",
        "parameters": [
          {
            "name": "featureId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LotMeasurement"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/features/search": {
      "get": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Search features",
        "description": "Search features by name or category",
        "operationId": "searchFeatures",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureMeta"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/features/code/{code}": {
      "get": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Get feature by code",
        "description": "Returns feature metadata by unique feature code (F0-F589)",
        "operationId": "getFeatureByCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureMeta"
                }
              }
            }
          }
        }
      }
    },
    "/features/category/{category}": {
      "get": {
        "tags": [
          "Feature Metadata"
        ],
        "summary": "Get features by category",
        "description": "Returns all features in a specific category (CVD_Process, Etch_Process, etc.)",
        "operationId": "getFeaturesByCategory",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeatureMeta"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "LocalTime": {
        "type": "object",
        "properties": {
          "hour": {
            "type": "integer",
            "format": "int32"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "second": {
            "type": "integer",
            "format": "int32"
          },
          "nano": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Shift": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "integer",
            "format": "int32"
          },
          "shiftCode": {
            "type": "string"
          },
          "shiftName": {
            "type": "string"
          },
          "startTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "endTime": {
            "$ref": "#/components/schemas/LocalTime"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Equipment": {
        "type": "object",
        "properties": {
          "equipmentId": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentCode": {
            "type": "string"
          },
          "equipmentName": {
            "type": "string"
          },
          "equipmentType": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "installDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Lot": {
        "type": "object",
        "properties": {
          "lotId": {
            "type": "integer",
            "format": "int32"
          },
          "lotNumber": {
            "type": "string"
          },
          "productType": {
            "$ref": "#/components/schemas/ProductType"
          },
          "equipment": {
            "$ref": "#/components/schemas/Equipment"
          },
          "operator": {
            "$ref": "#/components/schemas/Operator"
          },
          "shift": {
            "$ref": "#/components/schemas/Shift"
          },
          "productionStart": {
            "type": "string",
            "format": "date-time"
          },
          "productionEnd": {
            "type": "string",
            "format": "date-time"
          },
          "waferCount": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Operator": {
        "type": "object",
        "properties": {
          "operatorId": {
            "type": "integer",
            "format": "int32"
          },
          "operatorCode": {
            "type": "string"
          },
          "operatorName": {
            "type": "string"
          },
          "employeeNumber": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductType": {
        "type": "object",
        "properties": {
          "productTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "productCode": {
            "type": "string"
          },
          "productName": {
            "type": "string"
          },
          "productFamily": {
            "type": "string"
          },
          "targetYield": {
            "type": "number"
          },
          "specificationVersion": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "FeatureMeta": {
        "type": "object",
        "properties": {
          "featureId": {
            "type": "integer",
            "format": "int32"
          },
          "featureCode": {
            "type": "string"
          },
          "featureName": {
            "type": "string"
          },
          "featureCategory": {
            "type": "string"
          },
          "processStage": {
            "type": "string"
          },
          "measurementType": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "normalRangeMin": {
            "type": "number",
            "format": "double"
          },
          "normalRangeMax": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "isCritical": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LotMeasurement": {
        "type": "object",
        "properties": {
          "measurementId": {
            "type": "integer",
            "format": "int64"
          },
          "lot": {
            "$ref": "#/components/schemas/Lot"
          },
          "feature": {
            "$ref": "#/components/schemas/FeatureMeta"
          },
          "measurementValue": {
            "type": "number",
            "format": "double"
          },
          "isOutOfSpec": {
            "type": "boolean"
          },
          "measuredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageFeatureMeta": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeatureMeta"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}