# sam-api/src/main/java/com/secom/mes/entity/Lot.java
package com.secom.mes.entity;

import com.fasterxml.jackson.annotation.JsonIgnore;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;
import java.util.List;

@Entity
@Table(name = "lot")
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Lot {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "lot_id")
    private Integer lotId;

    @Column(name = "lot_number", nullable = false, unique = true, length = 100)
    private String lotNumber;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "product_type_id", nullable = false)
    private ProductType productType;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "equipment_id", nullable = false)
    private Equipment equipment;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "operator_id", nullable = false)
    private Operator operator;

    @ManyToOne(fetch = FetchType.LAZY)
    @JoinColumn(name = "shift_id", nullable = false)
    private Shift shift;

    @Column(name = "production_start", nullable = false)
    private LocalDateTime productionStart;

    @Column(name = "production_end")
    private LocalDateTime productionEnd;

    @Column(name = "wafer_count")
    private Integer waferCount;

    @Column(name = "status", length = 20)
    private String status; // 'in_progress', 'completed', 'quality_hold', 'released', 'scrapped'

    @Column(name = "created_at", updatable = false)
    private LocalDateTime createdAt;

    @Column(name = "updated_at")
    private LocalDateTime updatedAt;

    @OneToMany(mappedBy = "lot", cascade = CascadeType.ALL)
    @JsonIgnore
    private List<LotMeasurement> measurements;

    @OneToOne(mappedBy = "lot", cascade = CascadeType.ALL)
    @JsonIgnore
    private QualityResult qualityResult;

    @PrePersist
    protected void onCreate() {
        createdAt = LocalDateTime.now();
        updatedAt = LocalDateTime.now();
        if (waferCount == null) {
            waferCount = 25;
        }
    }

    @PreUpdate
    protected void onUpdate() {
        updatedAt = LocalDateTime.now();
    }
}